requirements.txt
mysql-connector-python==8.0.27
tkinter
tkcalendar==1.6.1
